Changes to the structural plasticity example to make it more understandable. Added some documentation and moved from examples/nest to pynest/examples
